.tool/lint: Ignore dupl complaints for cmd/oci-*/get.go
Don't worry about:

  $ make lint
  checking lint
  cmd/oci-cas/get.go:43::warning: duplicate of cmd/oci-refs/get.go:42-61 (dupl)
  cmd/oci-refs/get.go:42::warning: duplicate of cmd/oci-cas/get.go:43-62 (dupl)
  make: *** [lint] Error 1

The commands are all similar (open an engine, perform some method,
print the result), but are short enough that extracting out helpers
would be more trouble and indirection than it's worth.

Oddly, dupl seems happy to print:

  "duplicate of oci-cas/get.go:..."

and

  "duplicate of get.go:..."

if I exclude:

  "duplicate of cmd/oci-cas/get.go:..."

or

  "duplicate of .*oci-cas/get.go:..."

I want to get "oci-cas" and "oci-refs" in the exclusion regular
expression somewhere to avoid accidentally skipping dupl checks for
other get.go and similar if they show up somewhere else in the
repository, so I'm matching on the initial filename.

Signed-off-by: W. Trevor King <wking@tremily.us>