image: Add image-layout directory based CAS and ref engines
These are much nicer than the tar engines (hooray atomic renames :),
so switch the manifest tests tests back to using the directory-backed
engines.  I also switched the man-page examples over to
directory-backed layouts, now that they're what oci-image-init
generates by default.  And I added command-line wrappers for the
delete methods, now that we have a backend that implements it.

I do with there was a paginated, callback-based directory lister we
could use instead of ioutils.ReadDir.  On the other hand, by the time
directories get big enough for that to matter we may be sharding them
anyway.

Signed-off-by: W. Trevor King <wking@tremily.us>